requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.12.4,<4.0
python-dotenv>=0.21.0
requests>=2.28.0